Add external data preview to Codebook
