+++ src/components/Codebook/ExternalEntity.js
@@ -1,6 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import { Button } from '@codaco/ui';
 import useVariablesFromExternalData from '@hooks/useVariablesFromExternalData';
+import useExternalDataPreview from '@components/AssetBrowser/useExternalDataPreview';
 import VariableList from './VariableList';
 import EntityIcon from './EntityIcon';
 
@@ -10,28 +12,42 @@ const ExternalEntity = ({
 }) => {
   const { variables } = useVariablesFromExternalData(id);
 
+  const [preview, handleShowPreview] = useExternalDataPreview();
+
   return (
-    <div className="codebook__entity">
-      <div className="codebook__entity-detail">
-        <div className="codebook__entity-icon">
-          <EntityIcon entity="asset" />
+    <>
+      <div className="codebook__entity">
+        <div className="codebook__entity-detail">
+          <div className="codebook__entity-icon">
+            <EntityIcon entity="asset" />
+          </div>
+          <div className="codebook__entity-name">
+            <h2>
+              {name}
+            </h2>
+          </div>
+          <div className="codebook__entity-meta" />
+          <div className="codebook__entity-control">
+            <Button
+              size="small"
+              onClick={() => handleShowPreview(id)}
+            >
+              Preview asset
+            </Button>
+          </div>
         </div>
-        <div className="codebook__entity-name">
-          <h2>
-            {name}
-          </h2>
-        </div>
+        { variables.length > 0
+          && (
+          <div className="codebook__entity-variables">
+            <h3>Variables:</h3>
+            <VariableList
+              variables={variables}
+            />
+          </div>
+          )}
       </div>
-      { variables.length > 0
-        && (
-        <div className="codebook__entity-variables">
-          <h3>Variables:</h3>
-          <VariableList
-            variables={variables}
-          />
-        </div>
-        )}
-    </div>
+      {preview}
+    </>
   );
 };
